requirements.txt
pillow
pystray
PyGObject
configparser